backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pydantic-settings==2.5.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
redis==5.2.1
langgraph==0.2.20
langchain-groq==0.2.0
langchain-core==0.3.10
websockets==15.0.1
python-dotenv==1.0.1
structlog==24.4.0
